Rename internal/bundling into internal/building
