+++ src/internal/building/bundleToCompilationResult.js
