+++ src/internal/building/compositeAsset.js
