+++ src/internal/building/html/minifyHtml.js
