+++ src/internal/building/html/parseHtmlAsset.js
