+++ src/internal/building/html/parseSvgAsset.js
