+++ src/internal/building/isBareSpecifierForNativeNodeModule.js
