+++ src/internal/building/js/parseJsAsset.js
